ENH: database fixes

+++ sql/mysql/cdash-upgrade-1.4-1.6.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `coveragefilepriority` (
-  `id` bigint(20) NOT NULL,
+  `id` serial NOT NULL,
   `priority` tinyint(4) NOT NULL,
   `fullpath` varchar(255) NOT NULL,
   `projectid` int(11) NOT NULL,
