+++ sql/pgsql/cdash-upgrade-1.4-1.6.sql
@@ -2,7 +2,7 @@
 -- Table: coveragefilepriority
 --
 CREATE TABLE "coveragefilepriority" (
-  "id" bigint NOT NULL,
+  "id" serial NOT NULL,
   "priority" smallint NOT NULL,
   "fullpath" character varying(255) DEFAULT '' NOT NULL,
   "projectid" bigint  NOT NULL,
